feat(apigen): IN-FIX-TTL-TYPE unify TTL Go type to *uint32 in the catalog
Changed TTL GoType from goTypeInt64 to goTypeUint32 in the catalog for
ARecord, AAAARecord, CNAMERecord, MXRecord, and SRVRecord, and replaced
the literal "*uint32" in host_record.go with the goTypeUint32 constant.
Every TTL field in the DNS record catalog is now goTypeUint32 (aligning
with tools/openapi/inventory.md, which already documented ttl as *uint32
for all these resources).

Regenerated apis/{common,cluster,namespaced} and package/crds for the 5
affected kinds (10 CRD files: cluster + namespaced). CRDs now emit
format: int32 for ttl, preserving the existing Minimum=0/Maximum=2147483647
kubebuilder markers. Updated generator golden files for ARecord.

Added an Upgrade Notes section to README.md documenting the CRD schema
format change and the required CRD reapply on upgrade.

This catalog-level change makes the TTL field type *uint32 in the
generated apis types, which is a breaking change for the 5 controllers
(recorda, recordaaaa, recordcname, recordmx, recordsrv) that currently
plumb TTL as *int64 with a conversion helper (ttlOrZero/uint32PtrOrZero).
Removing those helpers and updating the controllers to use *uint32
directly is out of apigen's charter (internal/controller/** belongs to
the controller ant) and is tracked in a follow-up ticket on this same
branch. V.C52 (use-flag gate) was re-verified and still passes — this
change does not touch controller logic.

+++ apis/common/recordaaaa/recordaaaa_types.go
@@ -281,7 +281,7 @@ type AAAARecordParameters struct {
 	// Time-to-live in seconds. Zero means the record is not cached. Must be non-negative (0-2147483647); to inherit the zone/grid default, set useTtl to false rather than passing a negative sentinel value.
 	// +kubebuilder:validation:Minimum=0
 	// +kubebuilder:validation:Maximum=2147483647
-	TTL *int64 `json:"ttl,omitempty"`
+	TTL *uint32 `json:"ttl,omitempty"`
 	// Use flag for ttl — when false the zone/grid default TTL applies.
 	UseTTL *bool `json:"useTtl,omitempty"`
 	// Extensible attributes (arbitrary key/value metadata defined in Grid Manager). The WAPI wire format wraps each value as {"value": ...}; this map is the simplified string-valued CRD representation (the controller translates to/from the SDK's EA map[string]interface{} type).
@@ -317,7 +317,7 @@ type AAAARecordObservation struct {
 	// Comment for the record; maximum 256 characters.
 	Comment *string `json:"comment,omitempty"` // atProvider
 	// Time-to-live in seconds. Zero means the record is not cached. Must be non-negative (0-2147483647); to inherit the zone/grid default, set useTtl to false rather than passing a negative sentinel value.
-	TTL *int64 `json:"ttl,omitempty"` // atProvider
+	TTL *uint32 `json:"ttl,omitempty"` // atProvider
 	// Use flag for ttl — when false the zone/grid default TTL applies.
 	UseTTL *bool `json:"useTtl,omitempty"` // atProvider
 	// Extensible attributes (arbitrary key/value metadata defined in Grid Manager). The WAPI wire format wraps each value as {"value": ...}; this map is the simplified string-valued CRD representation (the controller translates to/from the SDK's EA map[string]interface{} type).
